Fix lint issues. Improve Insights package structure
