+++ Sources/InstantSearchInsights/Logic/EventProcessor+AlgoliaClient.swift
