+++ Sources/InstantSearchInsights/Models/Package+InsightsEvent.swift
