+++ Sources/InstantSearchInsights/Models/Packager+InsightsEvent.swift
